A lot of CaptainKirk & Claws done for now
I finished my stuff from yesterday which was making CaptainKirk go down and up, I have it automated, but I also have it mapped to buttons along with my other commands, this will probably change later depending on how much automation we want. I also made it so the claws can open and close, which are also now mapped. Maybe we'll use encoders to set positions for CaptianKirk, but for right now it's timed movement in whatever direction.