Lots of stuff, read desc.
I added a pneumatic for the lift mechanism, as all the ideas use one to push the hatch off. Then I changed the drivetrain to Victors, which made a lot of changes in many places. I also added a default command for the lift, and changed motor values so they aren't full speed